WIP IIPBanningCriterion
